resource_allocation module renamed to milp - stands for Mixed Integer Linear Programming - thinking can put linear models in there too - they share applications sometimes, and repeating models isn't good practice - instead, there should be option to make vars integer or real   - at least, for certain models that can apply to both   - this includes resource allocation model
