+++ ormm/mathprog/__init__.py
@@ -1,1 +1,1 @@
-from ormm.mathprog.resource_allocation import ResourceAllocation, print_sol
+from ormm.mathprog.milp import ResourceAllocation, print_sol
